requirements.txt
flask
ultralytics
pillow
gunicorn
flask-cors
opencv-python 